Audio Feedback + Audio Slider.

+++ Assets/Scripts/Database/GameDataReader.cs
@@ -503,7 +503,7 @@ public class GameDataReader : MonoBehaviour
                 _ammunitionData[i].critChance = myAmmunitionList.Ammunition[i].criticalHitChance;
                 _ammunitionData[i].critDamage = myAmmunitionList.Ammunition[i].criticalHitDamage;
                 _ammunitionData[i].range = myAmmunitionList.Ammunition[i].range;
-                _ammunitionData[i].cost = myAmmunitionList.Ammunition[i].cost;
+                _ammunitionData[i].cost = (int)myAmmunitionList.Ammunition[i].cost;
 				_ammunitionData[i].rarity = myAmmunitionList.Ammunition[i].rarity;
             }
             else
@@ -515,7 +515,7 @@ public class GameDataReader : MonoBehaviour
                 _ammunitionData[i].critChance = myAmmunitionList.Ammunition[i].criticalHitChance * _ammunitionData[0].critChance;
                 _ammunitionData[i].critDamage = myAmmunitionList.Ammunition[i].criticalHitDamage * _ammunitionData[0].critDamage;
                 _ammunitionData[i].range = myAmmunitionList.Ammunition[i].range * _ammunitionData[0].range;
-                _ammunitionData[i].cost = myAmmunitionList.Ammunition[i].cost * _ammunitionData[0].cost;
+                _ammunitionData[i].cost = (int)myAmmunitionList.Ammunition[i].cost * _ammunitionData[0].cost;
 				_ammunitionData[i].rarity = myAmmunitionList.Ammunition[i].rarity;
 			}
         }
@@ -533,7 +533,7 @@ public class GameDataReader : MonoBehaviour
                 _barrelData[i].critChance = myBarrelList.Barrel[i].criticalHitChance;
                 _barrelData[i].critDamage = myBarrelList.Barrel[i].criticalHitDamage;
                 _barrelData[i].range = myBarrelList.Barrel[i].range;
-                _barrelData[i].cost = myBarrelList.Barrel[i].cost;
+                _barrelData[i].cost = (int)myBarrelList.Barrel[i].cost;
 				_barrelData[i].rarity = myBarrelList.Barrel[i].rarity;
 			}
             else
@@ -545,7 +545,7 @@ public class GameDataReader : MonoBehaviour
                 _barrelData[i].critChance = myBarrelList.Barrel[i].criticalHitChance * _barrelData[0].critChance;
                 _barrelData[i].critDamage = myBarrelList.Barrel[i].criticalHitDamage * _barrelData[0].critDamage;
                 _barrelData[i].range = myBarrelList.Barrel[i].range * _barrelData[0].range;
-                _barrelData[i].cost = myBarrelList.Barrel[i].cost * _barrelData[0].cost;
+                _barrelData[i].cost = (int)myBarrelList.Barrel[i].cost * _barrelData[0].cost;
 				_barrelData[i].rarity = myBarrelList.Barrel[i].rarity;
 			}
         }
@@ -563,7 +563,7 @@ public class GameDataReader : MonoBehaviour
                 _gripData[i].critChance = myGripList.Grip[i].criticalHitChance;
                 _gripData[i].critDamage = myGripList.Grip[i].criticalHitDamage;
                 _gripData[i].range = myGripList.Grip[i].range;
-                _gripData[i].cost = myGripList.Grip[i].cost;
+                _gripData[i].cost = (int)myGripList.Grip[i].cost;
 				_gripData[i].rarity = myGripList.Grip[i].rarity;
 			}
             else
@@ -575,7 +575,7 @@ public class GameDataReader : MonoBehaviour
                 _gripData[i].critChance = myGripList.Grip[i].criticalHitChance * _gripData[0].critChance;
                 _gripData[i].critDamage = myGripList.Grip[i].criticalHitDamage * _gripData[0].critDamage;
                 _gripData[i].range = myGripList.Grip[i].range * _gripData[0].range;
-                _gripData[i].cost = myGripList.Grip[i].cost * _gripData[0].cost;
+                _gripData[i].cost = (int)myGripList.Grip[i].cost * _gripData[0].cost;
 				_gripData[i].rarity = myGripList.Grip[i].rarity;
 			}
         }
@@ -593,7 +593,7 @@ public class GameDataReader : MonoBehaviour
                 _magazineData[i].critChance = myMagazineList.Magazine[i].criticalHitChance;
                 _magazineData[i].critDamage = myMagazineList.Magazine[i].criticalHitDamage;
                 _magazineData[i].range = myMagazineList.Magazine[i].range;
-                _magazineData[i].cost = myMagazineList.Magazine[i].cost;
+                _magazineData[i].cost = (int)myMagazineList.Magazine[i].cost;
 				_magazineData[i].rarity = myMagazineList.Magazine[i].rarity;
                 _magazineData[i].capacity = myMagazineList.Magazine[i].capacity;
             }
@@ -606,7 +606,7 @@ public class GameDataReader : MonoBehaviour
                 _magazineData[i].critChance = myMagazineList.Magazine[i].criticalHitChance * _magazineData[0].critChance;
                 _magazineData[i].critDamage = myMagazineList.Magazine[i].criticalHitDamage * _magazineData[0].critDamage;
                 _magazineData[i].range = myMagazineList.Magazine[i].range * _magazineData[0].range;
-                _magazineData[i].cost = myMagazineList.Magazine[i].cost * _magazineData[0].cost;
+                _magazineData[i].cost = (int)myMagazineList.Magazine[i].cost * _magazineData[0].cost;
                 _magazineData[i].capacity = myMagazineList.Magazine[i].capacity * _magazineData[0].capacity;
                 _magazineData[i].rarity = myMagazineList.Magazine[i].rarity;
 			}
@@ -625,7 +625,7 @@ public class GameDataReader : MonoBehaviour
                 _sightData[i].critChance = mySightList.Sight[i].criticalHitChance;
                 _sightData[i].critDamage = mySightList.Sight[i].criticalHitDamage;
                 _sightData[i].range = mySightList.Sight[i].range;
-                _sightData[i].cost = mySightList.Sight[i].cost;
+                _sightData[i].cost = (int)mySightList.Sight[i].cost;
 				_sightData[i].rarity = mySightList.Sight[i].rarity;
             }
             else
@@ -637,7 +637,7 @@ public class GameDataReader : MonoBehaviour
                 _sightData[i].critChance = mySightList.Sight[i].criticalHitChance * _sightData[0].critChance;
                 _sightData[i].critDamage = mySightList.Sight[i].criticalHitDamage * _sightData[0].critDamage;
                 _sightData[i].range = mySightList.Sight[i].range * _sightData[0].range;
-                _sightData[i].cost = mySightList.Sight[i].cost * _sightData[0].cost;
+                _sightData[i].cost = (int)mySightList.Sight[i].cost * _sightData[0].cost;
 				_sightData[i].rarity = mySightList.Sight[i].rarity;
 			}
         }
@@ -655,7 +655,7 @@ public class GameDataReader : MonoBehaviour
                 _triggerMechanismData[i].critChance = myTriggerMechanismList.TriggerMechanism[i].criticalHitChance;
                 _triggerMechanismData[i].critDamage = myTriggerMechanismList.TriggerMechanism[i].criticalHitDamage;
                 _triggerMechanismData[i].range = myTriggerMechanismList.TriggerMechanism[i].range;
-                _triggerMechanismData[i].cost = myTriggerMechanismList.TriggerMechanism[i].cost;
+                _triggerMechanismData[i].cost = (int)myTriggerMechanismList.TriggerMechanism[i].cost;
 				_triggerMechanismData[i].rarity = myTriggerMechanismList.TriggerMechanism[i].rarity;
 			}
             else
@@ -667,7 +667,7 @@ public class GameDataReader : MonoBehaviour
                 _triggerMechanismData[i].critChance = myTriggerMechanismList.TriggerMechanism[i].criticalHitChance * _triggerMechanismData[0].critChance;
                 _triggerMechanismData[i].critDamage = myTriggerMechanismList.TriggerMechanism[i].criticalHitDamage * _triggerMechanismData[0].critDamage;
                 _triggerMechanismData[i].range = myTriggerMechanismList.TriggerMechanism[i].range * _triggerMechanismData[0].range;
-                _triggerMechanismData[i].cost = myTriggerMechanismList.TriggerMechanism[i].cost * _triggerMechanismData[0].cost;
+                _triggerMechanismData[i].cost = (int)myTriggerMechanismList.TriggerMechanism[i].cost * _triggerMechanismData[0].cost;
 				_triggerMechanismData[i].rarity = myTriggerMechanismList.TriggerMechanism[i].rarity;
 			}
 		}
